WIP. Dealing with docker socket

+++ deploy.py
@@ -2,10 +2,11 @@
 
 # System
 import sys
-import subprocess
+from subprocess import call
 import string
 import os
-import time
+from distutils.version import StrictVersion
+import json
 
 # Third Party
 import configargparse
@@ -27,23 +28,27 @@ import configargparse
 #                       [--rewrite-inspect]
 
 
-class Deployer:
+class Installer:
+
+    FLAVOR_VANILLA = "vanilla"
+    FLAVOR_DCOS = "dcos"
 
     def main(self):
 
         # Handle arguments
         self.parse_arguments()
         self.validate_arguments()
+        self.default_arguments()
         self.process_arguments()
 
         # Do the deed
-        self.deploy()
+        self.install()
 
 
     def parse_arguments(self):
 
         # Create an argument parser
-        self.parser = configargparse.ArgumentParser(description='Deploy Weave to a Mesos cluster on CentOS 7')
+        self.parser = configargparse.ArgumentParser(description='Install Weave to a Mesos cluster')
 
         # Add arguments to the parser
         self.add_mesos_arguments()
@@ -57,6 +62,26 @@ class Deployer:
 
         mesos_group = self.parser.add_argument_group('mesos', 'Mesos')
 
+        # Version
+        mesos_group.add_argument(
+            "--mesos-version",
+            dest="mesos_version",
+            env_var='MESOS_VERSION',
+            default="0.23.0",
+            help="Version of Mesos on slaves. Used to make decisions about how to install Weave. Default will probably work. (default: '%(default)s')"
+        )
+
+        # Flavor
+        # TODO: Anyone have a better name for this?
+        mesos_group.add_argument(
+            "--mesos-flavor",
+            dest="mesos_flavor",
+            env_var='MESOS_FLAVOR',
+            choices=[Installer.FLAVOR_VANILLA, Installer.FLAVOR_DCOS],
+            default=Installer.FLAVOR_VANILLA,
+            help="The 'flavor' of Mesos to install into. Determines several default values. (default: '%(default)s')"
+        )
+
         # Nodes
         mesos_group.add_argument(
             "--mesos-public-slaves",
@@ -64,7 +89,8 @@ class Deployer:
             env_var='MESOS_PUBLIC_SLAVES',
             nargs='*',
             type=str,
-            help="Space-separated list of public Mesos slave nodes names"
+            default=[],
+            help="Space-separated list of addresses of public Mesos slave nodes"
         )
         mesos_group.add_argument(
             "--mesos-private-slaves",
@@ -72,25 +98,16 @@ class Deployer:
             env_var='MESOS_PRIVATE_SLAVES',
             nargs='*',
             type=str,
-            help="Space-separated list of private Mesos slave nodes names"
+            default=[],
+            help="Space-separated list of addresses of private Mesos slave nodes"
         )
 
         # Admin username
         mesos_group.add_argument(
-            "--mesos-admin",
-            dest="mesos_admin",
-            env_var='MESOS_ADMIN',
-            default="core",
-            help="Admin username for Mesos nodes. (default: '%(default)s')"
-        )
-
-        # Configuration file
-        mesos_group.add_argument(
-            "--mesos-slave-config-file",
-            dest="mesos_slave_config_file",
-            env_var='MESOS_SLAVE_CONFIG_FILE',
-            default="/etc/default/mesos-slave",
-            help="Configuration file for Mesos slaves. (default: %(default)s)"
+            "--mesos-admin-username",
+            dest="mesos_admin_username",
+            env_var='MESOS_ADMIN_USERNAME',
+            help="Admin username for Mesos nodes. (default: Determined by 'flavor')"
         )
 
         # Service files
@@ -98,15 +115,13 @@ class Deployer:
             "--mesos-slave-service-file-public",
             dest="mesos_slave_service_file_public",
             env_var='MESOS_SLAVE_SERVICE_FILE_PUBLIC',
-            default="????",  # TODO: Find what a vanilla Mesos install uses
-            help="Public slave service file. TODO: Determine default. (default: %(default)s)"
+            help="Public slave systemd service file. (default: Determined by 'flavor')"
         )
         mesos_group.add_argument(
             "--mesos-slave-service-file-private",
             dest="mesos_slave_service_file_private",
             env_var='MESOS_SLAVE_SERVICE_FILE_PRIVATE',
-            default="????",  # TODO: Find what a vanilla Mesos install uses
-            help="Private slave service file. TODO: Determine default. (default: %(default)s)"
+            help="Private slave systemd service file. (default: Determined by 'flavor')"
         )
 
         # Service names
@@ -114,25 +129,13 @@ class Deployer:
             "--mesos-slave-service-name-public",
             dest="mesos_slave_service_name_public",
             env_var='MESOS_SLAVE_SERVICE_NAME_PUBLIC',
-            default="????",  # TODO: Find what a vanilla Mesos install uses
-            help="Name of Mesos public slave service. TODO: Determine default. (default: %(default)s)"
+            help="Name of Mesos public slave systemd service. (default: Determined by 'flavor')"
         )
         mesos_group.add_argument(
             "--mesos-slave-service-name-private",
             dest="mesos_slave_service_name_private",
             env_var='MESOS_SLAVE_SERVICE_NAME_PRIVATE',
-            default="????",  # TODO: Find what a vanilla Mesos install uses
-            help="Name of Mesos private slave service. TODO: Determine default. (default: %(default)s)"
-        )
-
-        # Hack
-        mesos_group.add_argument(
-            "--mesos-hack",
-            dest="mesos_hack",
-            env_var='MESOS_HACK',
-            action='store_true',
-            default=False,
-            help="HACK. (default: %(default)s)"
+            help="Name of Mesos private slave systemd service. (default: Determined by 'flavor')"
         )
 
 
@@ -146,7 +149,7 @@ class Deployer:
             dest="weave_install_dir",
             env_var='WEAVE_INSTALL_DIR',
             default=None,
-            help="The directory in which to install Weave. Must be on a writable volume. (default: /home/<mesos_admin>)"
+            help="The directory in which to install Weave. Must be on a writable volume. (default: /home/<mesos_admin_username>)"
         )
 
         # Components
@@ -163,7 +166,7 @@ class Deployer:
             "--weave-router-ipalloc-range",
             dest="weave_router_ipalloc_range",
             env_var='WEAVE_ROUTER_IPALLOC_RANGE',
-            default="10.20.0.0/16",
+            default="10.32.0.0/12",  # The Mesos default
             help="The range of IP numbers for Weave network nodes in CIDR form. (default: %(default)s)"
         )
 
@@ -198,15 +201,72 @@ class Deployer:
         if (len(self.args.mesos_public_slaves) == 0 and len(self.args.mesos_private_slaves) == 0):
             raise ValueError("You must specify at least one Mesos slave node using --mesos-public-slaves or --mesos_private_slaves")
 
+        # Make sure that the given Mesos version string is well-formed
+        StrictVersion(self.args.mesos_version)
+
+        # Validate the Mesos "flavor"
+        if not self.is_valid_mesos_flavor(self.args.mesos_flavor):
+            raise ValueError("Invalid mesos-flavor: " + self.args.mesos_flavor)
+
+
+    def is_valid_mesos_flavor(self, name):
+        if name is Installer.FLAVOR_VANILLA:
+            return True
+        if name is Installer.FLAVOR_DCOS:
+            return True
+        return False
+
+
+    def default_arguments(self):
+
+        # Vanilla flavor
+        if self.args.mesos_flavor == Installer.FLAVOR_VANILLA:
+            self.default_arguments_vanilla()
+
+        # DCOS flavor
+        elif self.args.mesos_flavor == Installer.FLAVOR_DCOS:
+            self.default_arguments_dcos()
+
+
+    def default_arguments_vanilla(self):
+        raise Exception("Not yet implemented. Someone needs to determine the defaults for 'vanilla' Mesos flavor")
+        # if self.args.mesos_admin_username is None:
+        #     self.args.mesos_admin_username = "TBD"
+        # if self.args.mesos_slave_service_file_public is None:
+        #     self.args.mesos_slave_service_file_public = "TBD"
+        # if self.args.mesos_slave_service_file_private is None:
+        #     self.args.mesos_slave_service_file_private = "TBD"
+        # if self.args.mesos_slave_service_name_public is None:
+        #     self.args.mesos_slave_service_name_public = "TBD"
+        # if self.args.mesos_slave_service_name_private is None:
+        #     self.args.mesos_slave_service_name_private = "TBD"
+
+
+    def default_arguments_dcos(self):
+        if self.args.mesos_admin_username is None:
+            self.args.mesos_admin_username = "core"
+        if self.args.mesos_slave_service_file_public is None:
+            self.args.mesos_slave_service_file_public = "/etc/systemd/system/dcos-mesos-slave-public.service"
+        if self.args.mesos_slave_service_file_private is None:
+            self.args.mesos_slave_service_file_private = "/etc/systemd/system/dcos-mesos-slave.service"
+        if self.args.mesos_slave_service_name_public is None:
+            self.args.mesos_slave_service_name_public = "dcos-mesos-slave-public.service"
+        if self.args.mesos_slave_service_name_private is None:
+            self.args.mesos_slave_service_name_private = "dcos-mesos-slave.service"
+
 
     def process_arguments(self):
 
-        # Let Weave installation directory default to the home directory of the Mesos admin user
-        if self.args.weave_install_dir is None:
-            self.args.weave_install_dir = "/home/" + self.args.mesos_admin
-
+        # Build directory paths for use later
         self.weave_bin_dir = self.args.weave_install_dir + "/bin"
         self.weave_tmp_dir = self.args.weave_install_dir + "/tmp"
+
+        # Let Weave installation directory default to the home directory of the Mesos admin user
+        if self.args.weave_install_dir is None:
+            self.args.weave_install_dir = "/home/" + self.args.mesos_admin_username
+
+        # Parse Mesos semantic version string
+        self.mesos_version = StrictVersion(self.args.mesos_version)
         
         # Build parameter substitution maps
         weave_router_peers = ' '.join(self.args.mesos_private_slaves + self.args.mesos_public_slaves)
@@ -220,18 +280,18 @@ class Deployer:
         ]
 
 
-    def deploy(self):
+    def install(self):
 
-        # Deploy to public Mesos slaves
+        # Install to public Mesos slaves
         for slave in self.args.mesos_public_slaves:
-            self.deploy_slave(slave, is_public=True)
+            self.install_into_slave(slave, is_public=True)
 
-        # Deploy to private Mesos slaves
+        # Install to private Mesos slaves
         for slave in self.args.mesos_private_slaves:
-            self.deploy_slave(slave, is_public=False)
+            self.install_into_slave(slave, is_public=False)
 
 
-    def deploy_slave(self, slave, is_public=False):
+    def install_into_slave(self, slave, is_public=False):
 
         print "------------------------------------------------------------------"
         print "Installing Weave into Mesos slave: " + slave
@@ -261,47 +321,60 @@ class Deployer:
         self.execute(slave, "sudo systemctl start weave-router.service")
         self.execute(slave, "sudo systemctl stop weave-proxy.service")
         self.execute(slave, "sudo systemctl start weave-proxy.service")
-        
+
         # Install weave proxy socket
-        if self.args.mesos_hack:
-            
-            # In older versions of Mesos and DCOS, the MESOS_DOCKER_SOCKET variable does not get picked up, so we must add the
-            # socket proxy directly to the command line where the slave is started by the service.
-            # Sample line for DCOS:
-            #       ExecStart=/opt/mesosphere/packages/mesos--d43a8eb9946a5c1c5ec05fb21922a2fdf41775b2/sbin/mesos-slave
-            # Sample line for vanilla Mesos:
-            #       <add example>
-            self.execute(slave, "sudo sed -i 's/^ExecStart(.*)mesos-slave$/ExecStart\1mesos-slave --docker_socket = {}' {}".format(self.args.weave_proxy_socket, self.args.mesos_slave_service_file_public))
-        
-        else:
-            
-            # Add the MESOS_DOCKER_SOCKET option to the slave configuration file
+        if self.mesos_version < StrictVersion("0.25.0"):
+
+            # Older versions of Mesos seem to pick up neither the MESOS_DOCKER_SOCKET option, nor mesos-slave --docker_host,
+            # or --docker-socket command-line options. The following *does* work
+
+            if (self.args.flavor is Installer.FLAVOR_VANILLA):
+                slave_executor_env_file = "TBD"
+                raise Exception("Not yet implemented. Someone needs to determine the location of the Docker executor environment properties file for 'vanilla' Mesos flavor")
+            else:
+                slave_executor_env_file = "/opt/mesosphere/etc/mesos-executor-environment.json"
+            self.add_property_to_json_file(slave, slave_executor_env_file, "DOCKER_HOST", "unix:///var/run/weave/weave.sock")
+
+        else: # version >= 0.25.0
+
+            # The following allegedly works for newer versions of Mesos, but this has not been tested, yet.
+
+            if (is_public):
+                slave_service_file = self.args.mesos_slave_service_file_public
+            else:
+                slave_service_file = self.args.mesos_slave_service_file_private
             line = "MESOS_DOCKER_SOCKET=" + self.args.weave_proxy_socket
-            file = self.args.mesos_slave_config_file
-            self.add_line_to_file(slave, line, file)
-        
+            self.add_line_to_file(slave, line, slave_service_file)
+
+            # If the above does not work, the following might, on DCOS Mesos
+            # command = "sudo sed -i 's|^ExecStart\\(.*\\)mesos-slave$|ExecStart\\1mesos-slave --docker_host=unix:://{}|' {}".format(self.args.weave_proxy_socket, slave_service_file)
+            # self.execute(slave, command)
+
         # Restart the Mesos slave so it picks up the new configuration
-        # if is_public:
-        #     service_name = self.args.mesos_slave_service_name_public
-        # else:
-        #     service_name = self.args.mesos_slave_service_name_private
-        # self.execute(slave, "sudo systemctl daemon-reload")
-        # self.execute(slave, "sudo systemctl stop " + service_name)
-        # self.execute(slave, "sudo systemctl start " + service_name)
+        if is_public:
+            service_name = self.args.mesos_slave_service_name_public
+        else:
+            service_name = self.args.mesos_slave_service_name_private
+        self.execute(slave, "sudo systemctl daemon-reload")
+        self.execute(slave, "sudo systemctl stop " + service_name)
+        self.execute(slave, "sudo systemctl start " + service_name)
 
 
     # Helpers -----------------------------------------------
 
     def copy(self, host, src_file, dst_dir, **kwargs):
         filename = os.path.basename(src_file)
-        dst_file = self.args.mesos_admin + "@" + host + ":" + dst_dir + "/" + filename
-        print "scp " + src_file + " " + dst_file
+        dst_file = self.args.mesos_admin_username + "@" + host + ":" + dst_dir + "/" + filename
+        description = "scp " + src_file + " " + dst_file
+        print description
         do_substitute = 'substitutions' in kwargs
         substituted_file = "/tmp/" + filename
         if do_substitute:
             substitutions = kwargs['substitutions']
             src_file = self.substitute(src_file, substitutions, substituted_file)
-        subprocess.call(["scp", src_file, dst_file])
+        result = call(["scp", src_file, dst_file])
+        if result is not 0:
+            raise Exception("Copying to remote failed with code: " + str(result))
         if do_substitute:
             os.remove(substituted_file)
 
@@ -319,9 +392,12 @@ class Deployer:
 
 
     def execute(self, host, command):
-        admin_at_host = self.args.mesos_admin + "@" + host
-        print "ssh " + admin_at_host + " " + command
-        subprocess.call(["ssh", admin_at_host, command])
+        admin_at_host = self.args.mesos_admin_username + "@" + host
+        description = "ssh " + admin_at_host + " " + command
+        print description
+        result = call(["ssh", admin_at_host, command])
+        if result is not 0:
+            raise Exception("Remote execution failed with code: " + str(result))
 
 
     def add_line_to_file(self, host, line, file):
@@ -331,7 +407,18 @@ class Deployer:
         self.execute(host, "sudo grep -q -F '{0}' {1} || printf '{0}\n' | sudo tee -a {1}".format(line, file))
 
 
+    def add_property_to_json_file(self, host, file, key, value):
+        if not os.path.exists(file):
+            raise Exception("Could not find file: " + file)
+        with open(file, "r") as source:
+            properties = json.load(source)
+
+        properties = json.loads(file)
+        if not properties.contains("kkk"):
+            pass
+
+
 # Main entry point
 if __name__ == "__main__":
-    deployer = Deployer()
-    deployer.main()
+    installer = Installer()
+    installer.main()
